Updating index merge to new maven api

+++ archiva-modules/archiva-base/archiva-maven2-indexer/src/main/java/org/apache/archiva/indexer/maven/merger/DefaultIndexMerger.java
@@ -19,15 +19,19 @@ package org.apache.archiva.indexer.maven.merger;
  */
 
 import org.apache.archiva.common.utils.FileUtils;
+import org.apache.archiva.indexer.UnsupportedBaseContextException;
 import org.apache.archiva.indexer.merger.IndexMerger;
 import org.apache.archiva.indexer.merger.IndexMergerException;
 import org.apache.archiva.indexer.merger.IndexMergerRequest;
 import org.apache.archiva.indexer.merger.TemporaryGroupIndex;
+import org.apache.archiva.repository.RepositoryRegistry;
+import org.apache.archiva.repository.RepositoryType;
 import org.apache.commons.lang.time.StopWatch;
-import org.apache.maven.index.NexusIndexer;
+import org.apache.maven.index.Indexer;
+import org.apache.maven.index.context.ContextMemberProvider;
 import org.apache.maven.index.context.IndexCreator;
 import org.apache.maven.index.context.IndexingContext;
-import org.apache.maven.index.context.UnsupportedExistingLuceneIndexException;
+import org.apache.maven.index.context.StaticContextMemberProvider;
 import org.apache.maven.index.packer.IndexPacker;
 import org.apache.maven.index.packer.IndexPackingRequest;
 import org.slf4j.Logger;
@@ -41,7 +45,10 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.Collection;
 import java.util.List;
+import java.util.Objects;
+import java.util.Optional;
 import java.util.concurrent.CopyOnWriteArrayList;
+import java.util.stream.Collectors;
 
 /**
  * @author Olivier Lamy
@@ -52,22 +59,28 @@ public class DefaultIndexMerger
     implements IndexMerger
 {
 
+    @Inject
+    RepositoryRegistry repositoryRegistry;
+
     private Logger log = LoggerFactory.getLogger( getClass() );
 
-    private final NexusIndexer indexer;
 
     private final IndexPacker indexPacker;
+
+    private Indexer indexer;
 
     private final List<IndexCreator> indexCreators;
 
     private List<TemporaryGroupIndex> temporaryGroupIndexes = new CopyOnWriteArrayList<>();
 
+    private List<IndexingContext>  temporaryContextes = new CopyOnWriteArrayList<>(  );
+
     private List<String> runningGroups = new CopyOnWriteArrayList<>();
 
     @Inject
-    public DefaultIndexMerger( NexusIndexer nexusIndexer, IndexPacker indexPacker, List<IndexCreator> indexCreators )
+    public DefaultIndexMerger( Indexer indexer, IndexPacker indexPacker, List<IndexCreator> indexCreators )
     {
-        this.indexer = nexusIndexer;
+        this.indexer = indexer;
         this.indexPacker = indexPacker;
         this.indexCreators = indexCreators;
     }
@@ -97,25 +110,29 @@ public class DefaultIndexMerger
         try
         {
             Path indexLocation = mergedIndexDirectory.resolve( indexMergerRequest.getMergedIndexPath() );
-            IndexingContext indexingContext =
-                indexer.addIndexingContext( tempRepoId, tempRepoId, mergedIndexDirectory.toFile(), indexLocation.toFile(), null, null,
-                                            indexCreators );
 
-            for ( String repoId : indexMergerRequest.getRepositoriesIds() )
-            {
-                IndexingContext idxToMerge = indexer.getIndexingContexts().get( repoId );
-                if ( idxToMerge != null )
-                {
-                    indexingContext.merge( idxToMerge.getIndexDirectory() );
-                }
-            }
-
-            indexingContext.optimize();
+            List<IndexingContext> members = indexMergerRequest.getRepositoriesIds( ).stream( ).map( id ->
+                repositoryRegistry.getRepository( id ) ).filter( repo -> repo.getType().equals( RepositoryType.MAVEN ) )
+                .map( repo -> {
+                    try
+                    {
+                        return repo.getIndexingContext().getBaseContext( IndexingContext.class );
+                    }
+                    catch ( UnsupportedBaseContextException e )
+                    {
+                        return null;
+                        // Ignore
+                    }
+                } ).filter( Objects::nonNull ).collect( Collectors.toList() );
+            ContextMemberProvider memberProvider = new StaticContextMemberProvider(members);
+            IndexingContext mergedCtx = indexer.createMergedIndexingContext( tempRepoId, tempRepoId, mergedIndexDirectory.toFile(),
+                indexLocation.toFile(), true, memberProvider);
+            mergedCtx.optimize();
 
             if ( indexMergerRequest.isPackIndex() )
             {
-                IndexPackingRequest request = new IndexPackingRequest( indexingContext, //
-                                                                       indexingContext.acquireIndexSearcher().getIndexReader(), //
+                IndexPackingRequest request = new IndexPackingRequest( mergedCtx, //
+                                                                       mergedCtx.acquireIndexSearcher().getIndexReader(), //
                                                                        indexLocation.toFile() );
                 indexPacker.packIndex( request );
             }
@@ -124,13 +141,14 @@ public class DefaultIndexMerger
             {
                 temporaryGroupIndexes.add( new TemporaryGroupIndex( mergedIndexDirectory, tempRepoId, groupId,
                                                                     indexMergerRequest.getMergedIndexTtl() ) );
+                temporaryContextes.add(mergedCtx);
             }
             stopWatch.stop();
             log.info( "merged index for repos {} in {} s", indexMergerRequest.getRepositoriesIds(),
                       stopWatch.getTime() );
-            return indexingContext;
+            return mergedCtx;
         }
-        catch ( IOException | UnsupportedExistingLuceneIndexException e )
+        catch ( IOException e)
         {
             throw new IndexMergerException( e.getMessage(), e );
         }
@@ -151,17 +169,19 @@ public class DefaultIndexMerger
 
         try
         {
-            IndexingContext indexingContext = indexer.getIndexingContexts().get( temporaryGroupIndex.getIndexId() );
-            if ( indexingContext != null )
-            {
-                indexer.removeIndexingContext( indexingContext, true );
+
+            Optional<IndexingContext> ctxOpt = temporaryContextes.stream( ).filter( ctx -> ctx.getId( ).equals( temporaryGroupIndex.getIndexId( ) ) ).findFirst( );
+            if (ctxOpt.isPresent()) {
+                IndexingContext ctx = ctxOpt.get();
+                indexer.closeIndexingContext( ctx, true );
+                temporaryGroupIndexes.remove( temporaryGroupIndex );
+                temporaryContextes.remove( ctx );
+                Path directory = temporaryGroupIndex.getDirectory();
+                if ( directory != null && Files.exists(directory) )
+                {
+                    FileUtils.deleteDirectory( directory );
+                }
             }
-            Path directory = temporaryGroupIndex.getDirectory();
-            if ( directory != null && Files.exists(directory) )
-            {
-                FileUtils.deleteDirectory( directory );
-            }
-            temporaryGroupIndexes.remove( temporaryGroupIndex );
         }
         catch ( IOException e )
         {
